implementing calendar formatting in the chart x-axis

+++ net.tourbook.chart/src/net/tourbook/chart/ChartMarkerLayer.java
@@ -72,8 +72,8 @@ public class ChartMarkerLayer implements IChartLayer {
 
 		final float graphYBottom = drawingData.getGraphYBottom();
 		final float[] yValues = drawingData.getYData().getHighValues()[0];
-		final float scaleX = drawingData.getScaleX();
-		final float scaleY = drawingData.getScaleY();
+		final double scaleX = drawingData.getScaleX();
+		final double scaleY = drawingData.getScaleY();
 
 		final Color colorLine = new Color(display, getLineColor());
